radiomanager-frontend: Rename and restructure API module
