+++ services/radiomanager-frontend/src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import cn from 'classnames'
-import { User } from '@/api/api.types'
+import { User } from '@/api/apiTypes'
 import Local from 'next/font/local'
 import Image from 'next/image'
 import Logo from './logo.png'
