+++ services/radiomanager-frontend/src/components/LibraryTracksList/LibraryTracksList.tsx
@@ -1,4 +1,4 @@
-import { UserTrack } from '@/api/api.types'
+import { UserTrack } from '@/api/apiTypes'
 import { TracksList } from '@/components/common/TrackList'
 import { InfiniteScroll } from '@/components/common/InfiniteScroll/InfiniteScroll'
 import AnimatedBars from '@/icons/AnimatedBars'
