+++ services/radiomanager-frontend/src/components/common/TrackList/helpers.ts
@@ -1,6 +1,5 @@
-export const isMac = window.navigator.userAgent.includes('Mac')
-
 export const isModifierKeyPressed = (event: React.MouseEvent) => {
+  const isMac = window.navigator.userAgent.includes('Mac')
   const key = isMac ? 'metaKey' : 'ctrlKey'
 
   return event[key]
