+++ services/radiomanager-frontend/src/components/entries/ChannelPage/hooks/useHandleChannelLastUploadedTrack.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { useMediaUploader } from '@/modules/MediaUploader'
 import { UploadedTrackType } from '@/modules/MediaUploader/MediaUploaderTypes'
-import { UserChannelTrack } from '@/api/api.types'
+import { UserChannelTrack } from '@/api/apiTypes'
 
 export const useHandleChannelLastUploadedTrack = (
   channelId: number,
